refactor(store): split supersede-terminal-spawn marking into RecordingStore.SupersedeTerminalSpawn.cs
